Add an option to load all gists asynchrously
Until now, all gist will be load via https://gist.github.com/$id.js which is only a script containing 2 document.write(!). Considering website performance (document.write is a blocker), this change introduce deferred loading of gists. At the moment, this is completely backwards compatible and is *not turned on at default*, because it requires jQuery (jQuery is not included maybe).

To enable, just define `EMBED_GISTHUB_DEFERRED` to true.